moved apply applicant tests to listings
