+++ tests/services/test_listing_service.py
@@ -2,7 +2,9 @@ import pytest
 from unittest.mock import patch, Mock
 from chalicelib.services.ListingService import ListingService
 from chalice.app import NotFoundError
-
+from chalice.app import Response, NotFoundError
+from datetime import datetime, timezone
+import uuid
 
 SAMPLE_LISTINGS = [
     {
@@ -34,17 +36,20 @@ SAMPLE_LISTINGS = [
 @pytest.fixture
 def service():
     with patch("chalicelib.services.ListingService.RepositoryFactory") as mock_factory:
+        mock_applicants_repo = Mock()
         mock_listings_repo = Mock()
+        mock_factory.applications.return_value = mock_applicants_repo
         mock_factory.listings.return_value = mock_listings_repo
 
-        service = ListingService()  # will use the patched RepositoryFactory
-        yield service, mock_listings_repo
+        # services: will use the patched RepositoryFactory
+        listing_service = ListingService()
+        yield listing_service, mock_listings_repo, mock_applicants_repo
 
 
 def test_create_listing(service):
     import uuid
 
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     CREATED_LISTING = {
         "title": "test",
@@ -71,7 +76,7 @@ def test_create_listing(service):
 
 
 def test_get_listing(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.get_by_id.return_value = SAMPLE_LISTINGS[0]
 
@@ -84,7 +89,7 @@ def test_get_listing(service):
 
 
 def test_get_all_listings(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.get_all.return_value = SAMPLE_LISTINGS
 
@@ -95,7 +100,7 @@ def test_get_all_listings(service):
 
 
 def test_delete_listing(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.delete.return_value = True
 
@@ -106,7 +111,7 @@ def test_delete_listing(service):
 
 
 def test_delete_listing_not_found(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.delete.side_effect = NotFoundError("Listing not found.")
 
@@ -115,7 +120,7 @@ def test_delete_listing_not_found(service):
 
 
 def test_toggle_visibility(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listing_id = SAMPLE_LISTINGS[0]["id"]
     result = listing_service.toggle_visibility(id=mock_listing_id)
@@ -128,7 +133,7 @@ def test_toggle_visibility(service):
 
 
 def test_toggle_visibility_invalid_listing_id(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.toggle_boolean_field.return_value = None
     mock_listings_repo.toggle_boolean_field.side_effect = NotFoundError(
@@ -146,7 +151,7 @@ def test_toggle_visibility_invalid_listing_id(service):
 
 
 def test_toggle_visibility_exception(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.toggle_boolean_field.return_value = None
     mock_listings_repo.toggle_boolean_field.side_effect = Exception("Error.")
@@ -160,7 +165,7 @@ def test_toggle_visibility_exception(service):
 
 
 def test_update_field_route(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.update_field.return_value = None
     mock_listings_repo.update.return_value = None
@@ -178,7 +183,7 @@ def test_update_field_route(service):
 
 
 def test_update_field_listing_not_found(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.update_field.side_effect = NotFoundError("Listing not found.")
 
@@ -192,7 +197,7 @@ def test_update_field_listing_not_found(service):
 
 
 def test_update_field_exception(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.update_field.side_effect = Exception("Error.")
 
@@ -206,7 +211,7 @@ def test_update_field_exception(service):
 
 
 def test_toggle_encryption_succeeds(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
     mock_listings_repo.toggle_boolean_field.return_value = True
     mock_listing_id = SAMPLE_LISTINGS[0]["id"]
@@ -221,9 +226,11 @@ def test_toggle_encryption_succeeds(service):
 
 
 def test_toggle_encryption_invalid_listing_id_raises_not_found(service):
-    listing_service, mock_listings_repo = service
+    listing_service, mock_listings_repo, _ = service
 
-    mock_listings_repo.toggle_boolean_field.side_effect = NotFoundError("Listing not found: 3.")
+    mock_listings_repo.toggle_boolean_field.side_effect = NotFoundError(
+        "Listing not found: 3."
+    )
     mock_listing_id = "3"
 
     with pytest.raises(NotFoundError) as exc_info:
@@ -234,3 +241,127 @@ def test_toggle_encryption_invalid_listing_id_raises_not_found(service):
     )
 
     assert str(exc_info.value) == "Listing not found: 3."
+
+
+@patch(
+    "chalicelib.services.ListingService.uuid.uuid4",
+    return_value=uuid.UUID("12345678-1234-5678-1234-567812345678"),
+)
+@patch("chalicelib.services.ListingService.Application.model_validate")
+@patch(
+    "chalicelib.services.ListingService.get_file_extension_from_base64",
+    return_value="png",
+)
+@patch("chalicelib.services.ListingService.s3.upload_binary_data")
+@patch("chalicelib.services.ListingService.ses.send_email")
+@patch("chalicelib.services.ListingService.datetime")
+def test_apply_success(
+    mock_datetime,
+    mock_send_email,
+    mock_upload,
+    mock_get_ext,
+    mock_validate,
+    mock_uuid,
+    service,
+):
+    listing_service, mock_listings_repo, mock_applicants_repo = service
+
+    # Setup fixed current time and deadline
+    mock_datetime.now.return_value = datetime(2023, 9, 10, tzinfo=timezone.utc)
+    mock_datetime.fromisoformat.return_value = datetime(
+        2023, 9, 19, tzinfo=timezone.utc
+    )
+
+    # Mock listing from repo
+    mock_listings_repo.get_by_id.return_value = {
+        "deadline": "2023-09-19T04:00:00.000Z",
+        "is_visible": True,
+    }
+
+    # Mock S3 upload return values
+    mock_upload.side_effect = [
+        "https://s3/resume.pdf",
+        "https://s3/image.png",
+    ]
+
+    # Sample application input
+    application_data = {
+        "listing_id": "1",
+        "first_name": "Jane",
+        "last_name": "Doe",
+        "resume": b"resume-bytes",
+        "image": "base64-image-data",
+        "email": "jane@example.com",
+    }
+
+    result = listing_service.apply(application_data.copy())
+
+    # Validate calls and result
+    mock_validate.assert_called_once()
+    assert mock_upload.call_count == 2
+    assert mock_send_email.called
+    mock_applicants_repo.create.assert_called_once()
+
+    assert result == {
+        "msg": True,
+        "resumeUrl": "https://s3/resume.pdf",
+    }
+
+    # Also check that `resume` and `image` were replaced with URLs
+    uploaded_data = mock_applicants_repo.create.call_args[1]["data"]
+    assert uploaded_data["resume"] == "https://s3/resume.pdf"
+    assert uploaded_data["image"] == "https://s3/image.png"
+
+
+@patch("chalicelib.services.ListingService.Application.model_validate")
+def test_apply_listing_not_found(model_validate, service):
+    listing_service, mock_listings_repo, _ = service
+
+    # Mock invisible listing
+    mock_listings_repo.get_by_id.return_value = {
+        "is_visible": False,
+        "deadline": "2100-01-01T00:00:00.000Z",
+    }
+
+    with pytest.raises(NotFoundError, match="Invalid listing."):
+        listing_service.apply(
+            {
+                "listing_id": "1",
+                "last_name": "Doe",
+                "first_name": "John",
+                "resume": "fake_resume",
+                "image": "fake_image",
+                "email": "john@example.com",
+            }
+        )
+
+
+@patch("chalicelib.services.ListingService.datetime")
+@patch("chalicelib.services.ListingService.Application.model_validate")
+def test_apply_deadline_passed(mock_validate, mock_datetime, service):
+    listing_service, mock_listings_repo, _ = service
+
+    # Listing is visible but deadline is in the past
+    mock_listings_repo.get_by_id.return_value = {
+        "is_visible": True,
+        "deadline": "2000-01-01T00:00:00.000+00:00",
+    }
+
+    # Force "now" to be far in the future
+    mock_datetime.now.return_value = datetime(2025, 1, 1, tzinfo=timezone.utc)
+    mock_datetime.fromisoformat.side_effect = datetime.fromisoformat
+
+    response = listing_service.apply(
+        {
+            "listing_id": "1",
+            "last_name": "Doe",
+            "first_name": "John",
+            "resume": "fake_resume",
+            "image": "fake_image",
+            "email": "john@example.com",
+        }
+    )
+
+    assert isinstance(response, Response)
+    assert response.status_code == 410
+    assert "deadline" in response.body.lower()
